+++ hutool-poi/src/main/java/com/xiaoleilu/hutool/poi/excel/ExcelReader.java
@@ -21,31 +21,95 @@ public class ExcelReader {
 	
 	/** 是否忽略空行 */
 	private boolean ignoreEmptyRow;
-	/** 是否忽略空行 */
+	/** 是否去除单元格元素两边空格 */
+	private boolean trimCellValue;
+	/** 标题别名 */
 //	private Map<String, String> headerAlias;
 	private Sheet sheet;
 
 	// ------------------------------------------------------------------------------------------------------- Constructor start
+	/**
+	 * 构造
+	 * @param bookStream Excel文件的流
+	 * @param sheetIndex sheet序号，0表示第一个sheet
+	 */
 	public ExcelReader(InputStream bookStream, int sheetIndex) {
 		this(ExcelUtil.loadBook(bookStream), sheetIndex);
 	}
 
+	/**
+	 * 构造
+	 * @param bookStream Excel文件的流
+	 * @param sheetName sheet名，第一个默认是sheet1
+	 */
 	public ExcelReader(InputStream bookStream, String sheetName) {
 		this(ExcelUtil.loadBook(bookStream), sheetName);
 	}
 
+	/**
+	 * 构造
+	 * @param book {@link Workbook} 表示一个Excel文件
+	 * @param sheetIndex sheet序号，0表示第一个sheet
+	 */
 	public ExcelReader(Workbook book, int sheetIndex) {
 		this(book.getSheetAt(sheetIndex));
 	}
 
+	/**
+	 * 构造
+	 * @param book {@link Workbook} 表示一个Excel文件
+	 * @param sheetName sheet名，第一个默认是sheet1
+	 */
 	public ExcelReader(Workbook book, String sheetName) {
 		this(book.getSheet(sheetName));
 	}
 
+	/**
+	 * 构造
+	 * @param sheet Excel中的sheet
+	 */
 	public ExcelReader(Sheet sheet) {
 		this.sheet = sheet;
 	}
 	// ------------------------------------------------------------------------------------------------------- Constructor end
+	
+	// ------------------------------------------------------------------------------------------------------- Getters and Setters start
+	/**
+	 * 是否忽略空行
+	 * @return 是否忽略空行
+	 */
+	public boolean isIgnoreEmptyRow() {
+		return ignoreEmptyRow;
+	}
+
+	/**
+	 * 设置是否忽略空行
+	 * @param ignoreEmptyRow 是否忽略空行
+	 * @return this
+	 */
+	public ExcelReader setIgnoreEmptyRow(boolean ignoreEmptyRow) {
+		this.ignoreEmptyRow = ignoreEmptyRow;
+		return this;
+	}
+
+	/**
+	 * 是否去掉单元格值两边空格
+	 * @return 是否去掉单元格值两边空格
+	 */
+	public boolean isTrimCellValue() {
+		return trimCellValue;
+	}
+
+	/**
+	 * 设置是否去掉单元格值两边空格
+	 * @param trimCellValue 是否去掉单元格值两边空格
+	 * @return this
+	 */
+	public ExcelReader setTrimCellValue(boolean trimCellValue) {
+		this.trimCellValue = trimCellValue;
+		return this;
+	}
+	// ------------------------------------------------------------------------------------------------------- Getters and Setters end
 
 	/**
 	 * 读取工作簿中指定的Sheet的所有行列数据
@@ -53,7 +117,7 @@ public class ExcelReader {
 	 * @param startRowIndex 起始行
 	 * @return 行的集合，一行使用List表示
 	 */
-	public List<List<Object>> readAll() {
+	public List<List<Object>> read() {
 		return read(0, Integer.MAX_VALUE);
 	}
 
@@ -92,7 +156,7 @@ public class ExcelReader {
 			Cell cell;
 			while (celIter.hasNext()) {
 				cell = celIter.next();
-				cellValues.add(ExcelUtil.getCellValue(cell));
+				cellValues.add(ExcelUtil.getCellValue(cell, this.trimCellValue));
 			}
 		}
 		return cellValues;